Checksum Generator refactor to be OSGi service rather than static Util
